pyheap/pyproject.toml
[tool.poetry]
name = "pyheap"
version = "0.5.0"
description = ""
authors = ["Ivan Yurchenko <ivan@ivanyu.me>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
pyelftools = "^0.29"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
black = "^22.6"
pex = "^2.1"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = ["src"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
